SquareMenuLambda/getMenuObjects/requirements.txt
requests
urllib3<2
squareup